+++ components/site/tests/site.rs
@@ -1,6 +1,5 @@
 mod common;
 
-use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 
@@ -324,7 +323,7 @@ fn can_build_site_with_taxonomies() {
             let pages_data = std::mem::replace(&mut library.pages, AHashMap::new());
             for (i, (_, mut page)) in pages_data.into_iter().enumerate() {
                 page.meta.taxonomies = {
-                    let mut taxonomies = HashMap::new();
+                    let mut taxonomies = std::collections::BTreeMap::new();
                     taxonomies.insert(
                         "categories".to_string(),
                         vec![if i % 2 == 0 { "A" } else { "B" }.to_string()],
@@ -646,7 +645,7 @@ fn can_build_site_with_pagination_for_taxonomy() {
                     nb_a_pages += 1;
                 }
                 page.meta.taxonomies = {
-                    let mut taxonomies = HashMap::new();
+                    let mut taxonomies = std::collections::BTreeMap::new();
                     taxonomies.insert(
                         "tags".to_string(),
                         vec![if i % 2 == 0 { "A" } else { "B" }.to_string()],